Fix class def to match implementation.
git-svn-id: https://arducopter.googlecode.com/svn/trunk@505 f9c3cf11-9bcb-44bc-f272-b75c42450872

+++ libraries/Waypoints/Waypoints.h
@@ -18,7 +18,7 @@ class Waypoints
 		int32_t lng;				// Longitude * 10**7
 	};
 	
-	WP get_waypoint_with_index(uint16_t i);
+	WP get_waypoint_with_index(uint8_t i);
 	WP get_current_waypoint(void);
 	WP get_next_waypoint(void);
